finalize the tvl fallback
