+++ src/hooks/data/vaults/fallbacks/tvl/useVaultTotalAssets.ts
@@ -1,0 +1,27 @@
+import { useContractRead } from "wagmi"
+
+import { VaultDynamicProps } from "@/lib/types"
+import useActiveChainId from "@/hooks/useActiveChainId"
+
+// work also for the token compounder
+import { AMMCompounderBase } from "@/constant/abi"
+
+export default function useVaultTotalAssets({
+  vaultAddress,
+  enabled,
+}: {
+  vaultAddress: VaultDynamicProps["vaultAddress"]
+  enabled: boolean
+}) {
+  const chainId = useActiveChainId()
+
+  const totalAssetsQuery = useContractRead({
+    chainId,
+    abi: AMMCompounderBase,
+    address: vaultAddress,
+    functionName: "totalAssets",
+    enabled: enabled,
+  })
+
+  return totalAssetsQuery
+}
